### 2017-06-08 #283 @whedon
[PRE REVIEW]: Web-based text anonymization with Node.js: Introducing NETANOS (Named entity-based Text Anonymization for Open Science)
**Submitting author:** @ben-aaron188 (<a href="http://orcid.org/0000-0003-1658-9086">Bennett Kleinberg</a>)
**Repository:** <a href="https://github.com/ben-aaron188/netanos" target ="_blank">https://github.com/ben-aaron188/netanos</a>
**Version:** v1.1.0
