### 2017-06-08 #283 @arfon
:wave: thanks for your submission to JOSS. From a quick inspection of this submission it's not entirely obvious that it meets our [submission criteria](http://joss.theoj.org/about#author_guidelines). In particular, this item:

> - Your software should have a research application

Could you confirm here that there _is_ a research application for this software (and explain what that application is)?

Many thanks!
