### 2017-06-08 #283 @ben-aaron188
@arfon Hi. Yes, our tool is written for the purpose of helping researchers share their text data. In the linguistic and content-based research field, a key reason why researchers don't post their data online (and hence refuse access to others) is that the data needs to be anonymized in order to protect the privacy of the participants who provided (or are mentioned in) the text data (e.g interviews).
The research application is thus that our tool will be a step towards enabling researchers to finally share text data without the need of manual anonymization. Consequently, it will enable proper re-analyses, meta-analyses and other secondary analysis on data that otherwise is unaccessible to anybody other than the original researchers. Eventually, this could lead to a big boost towards open data in the text-based research community (i.e. comp. social science, psychology, linguistics).
Please do let me know if I can clarify this further.
