### 2017-06-08 #283 @acabunoc
Hi @richardlitt & @nuest 👋 would either of you be up for reviewing this?

Review guidelines can be found here: http://joss.theoj.org/about#reviewer_guidelines
