### 2017-06-11 ben-aaron188/netanos#1 @RichardLitt
Merge npm_release with master branch
I don't understand the rationale for two branches. Having the package.json in master would be useful for people wanting to find it (I had to search a bit to figure it out, especially as the URL on npm goes to a blob which isn't there). These two branches are also out of sync.

Related to https://github.com/openjournals/joss-reviews/issues/283#issuecomment-307141013.
