### 2017-06-11 ben-aaron188/netanos#2 @RichardLitt
Add tests
These should be relatively easy to set up, as the API is not complicated. Adding one for each of the main methods would be fantastic, and then making them runnable through an npm script.

Related to https://github.com/openjournals/joss-reviews/issues/283#issuecomment-307141013.
