### 2017-06-11 ben-aaron188/netanos#3 @RichardLitt
Add a CONTRIBUTE.md document
Currently, I don't know if I can contribute to this code at all. Are you interested in long term maintenance?

A Code of Conduct would also be good.

Related to https://github.com/openjournals/joss-reviews/issues/283#issuecomment-307141013.
