### 2017-06-11 #283 @RichardLitt
I've partially reviewed the paper. Due to my current work circumstances (not on my main machine and traveling), I wasn't able to install Java to test out the install. I'll be able to do this in approximately 14 days when I am back with my work computer. I've added a reminder to my calendar to come back to this review at that time.

In the meanwhile, I've raised three issues above which I feel ought to be met before this can be accepted. They're all relatively minor, I believe.
