### 2017-06-14 #283 @arfon
> I've partially reviewed the paper. Due to my current work circumstances (not on my main machine and traveling), I wasn't able to install Java to test out the install. I'll be able to do this in approximately 14 days when I am back with my work computer. I've added a reminder to my calendar to come back to this review at that time.

❤️ many thanks @RichardLitt. I'm going to move this over to an actual review issue so when you come back we can wrap this up.
