### 2017-06-14 #293 @whedon
[REVIEW]: Web-based text anonymization with Node.js: Introducing NETANOS (Named entity-based Text Anonymization for Open Science)
**Submitting author:** @ben-aaron188 (<a href="http://orcid.org/0000-0003-1658-9086">Bennett Kleinberg</a>)
**Repository:** <a href="https://github.com/ben-aaron188/netanos" target ="_blank">https://github.com/ben-aaron188/netanos</a>
**Version:** v1.1.5
