### 2017-06-15 27a77c7
bugfix: partial anonymization callstack
