### 2017-06-15 d1a7ba6
minor npm update
