### 2017-06-17 2ad9c8b
added mocha.js tests for Netanos.js interface
