### 2017-06-17 8e6efd1
added tests for ner replacement
