### 2017-06-18 bd94d28
upd README for API tests
