### 2017-06-18 ben-aaron188/netanos#3 @ben-aaron188
Thanks for pointing us to the CoC and contributing guidelines. We've added both.
