### 2017-06-18 ben-aaron188/netanos#2 @ben-aaron188
We've included tests for each of the main method, runnable through `npm test` after installing mochajs.
