### 2017-06-18 0ff17f5
Delete CONTRIBUTING.md
