### 2017-06-18 d5a4b81
Create CONTRIBUTING.md
