+++ CONTRIBUTING.md
@@ -1,0 +1,48 @@
+# Contributing to NETANOS
+
+Thanks for contributing to NETANOS.
+
+NETANOS is intended to grow over time to become the scienitific standard for text anonymization in the spirit of open code, open data, and open science.
+
+Here are some guidelines on how to best contribute:
+
+#### How can I contribute?
+
+We appreciate contributions of all kind (bug reports, suggested features, questions that are useful for all NETANOS users). Since not all users of NETANOS are coders, we also appreciate non-technical suggestions for improvements of the tool as well as ideas on how to make the web interface more accessible.
+
+The easiest way to contribute (and the fastest way to respond for us) is if you submit an issue on this repository. If you include the purpose of the issue (bug report, question, suggestion) and a clear description of the issue, that will help us get started. If it concerns a bug report, please add reproducible code of possible, console output, or a screenshot of the issue.
+
+##### For coders:
+
+We have tried to make the source code as readable and accessible as possible. If you use the structure of code that we wrote NETANOS in, we look forward to any pull requests :-)
+
+
+
+#### Wish list of features
+
+Some future developments that we think would be great additions to NETANOS are:
+
+- Evaluation of the "anonymizability" of input (e.g. how suited is the text to be well anonymised?)
+- Different named entity recognisers as dependencies (e.g. [Spacy's NER](https://github.com/explosion/spaCy), [Google Cloud Natural Language API](https://cloud.google.com/natural-language/docs/basics#entity-analysis-response)
+- Semantically consistent named entity replacements (e.g. `"I love the captial of Germany. Berlin is a great startup hub."` becomes `"I love the captial of Spain. Madrid is a great startup hub."`)
+
+We are currently working on a release of the NETANOS web interface for simple drag-and-drop functionality. Keep an eye on the repo for the link and release notes.
+
+
+
+#### What's the Code of Conduct?
+
+Check our Code of Conduct [here](https://github.com/ben-aaron188/netanos/blob/master/CODE_OF_CONDUCT.md).
+
+
+
+#### Can I contact the maintainers directly?
+
+Yes. In general, however, we believe it's best if all issues, requests and questions are discussed directly on the netanos repo on GitHub. If you do feel that something is best communicated directly, contact us at projectnetanos@gmail.com
+
+
+
+
+
+
+
