### 2017-06-22 #293 @RichardLitt
Version: The GitHub release does not match, and the npm release is now ahead of the release for this review (as it is 1.1.3, instead of 1.1.0). I don't know how to best suggest moving forward; perhaps committing and tagging a GitHub release as 1.1.3, and then updating this submission here?
