### 2017-06-22 ben-aaron188/netanos#4 @RichardLitt
Add DOIs to references in paper.md
In https://github.com/ben-aaron188/netanos/blob/master/paper/paper.md, you have some references. Each should have a DOI to point to, as well. If there aren't any, state that here and close this.

Related to https://github.com/openjournals/joss-reviews/issues/293.
