### 2017-06-22 ben-aaron188/netanos#5 @RichardLitt
Add ner and promise to package.json dependencies
Instead of specifying in the README which modules the user needs to download, run this command:

```sh
npm install --save ner
npm install --save promise
```

And commit the changed package.json file. This way, in the future, the user can just run `npm i`.

Related to https://github.com/openjournals/joss-reviews/issues/293.
