### 2017-06-22 ben-aaron188/netanos#6 @RichardLitt
Use `.` instead of `$scriptDir` in Java process 
In the README, you suggest using `$scriptDir` as a variable for where the stanford NER parser was downloaded. However, I found it easier to just `cd` into that directory, and run this:

```
java -mx1000m -cp "./stanford-ner.jar:./lib/*" edu.stanford.nlp.ie.NERServer  -loadClassifier ./classifiers/english.muc.7class.distsim.crf.ser.gz -port 8080 -outputFormat inlineXML
```

I think using `.` is a bit cleaner and easier. Just an idea, this isn't a mandatory change for https://github.com/openjournals/joss-reviews/issues/293.
