### 2017-06-22 #293 @RichardLitt
Opened issues for Version and References. Otherwise, this is all good. Note: not sure what to do about versions. See above.
