### 2017-06-22 #293 @RichardLitt
Note: I am happy with the changes I suggested in https://github.com/openjournals/joss-reviews/issues/283. These have now been fixed.
