### 2017-06-23 e7be4b6
added dependencies to package.json; added DOIs to references; adjusted README.md
