### 2017-06-24 f1f852c
minimal adj in Readme
