### 2017-06-24 ben-aaron188/netanos#4 @ben-aaron188
We added the DOIs where they were available. For "NLP compromise",  "ner" and the UK Data Archive anonymization guideline, there weren't any.
