### 2017-06-24 #293 @ben-aaron188
@RichardLitt Thanks for pointing that versioning issue out. We've tried to address this as follows:

1. Latest version is in the GitHub repo
2. Latest version is in the npm release with v1.1.4
3. On the GitHub repo, we created a release tag to the master branch with v1.1.4 and refer to the npm link in the release description

Does this sound like an option? We would need to change the version for the JOSS submission then though.
