### 2017-06-24 #293 @RichardLitt
Sounds good to me! Changing the version in the JOSS submission also seems like a smart move. @arfon that's the last thing that needs to be done, and then I consider this to be worthy of accepting as is.
