### 2017-06-24 #293 @RichardLitt
@arfon Looks like we need to have an 'Update version' command, and an 'accept paper' command.
