### 2017-06-25 #293 @arfon
> @arfon Looks like we need to have an 'Update version' command, and an 'accept paper' command.

👍 yep, that's next on my todo list.
