### 2017-06-25 #293 @arfon
@ben-aaron188 - Could you move the references you currently have in the `paper.md` file into a `paper.bib` file and cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))

I've also made a small fix to your paper meta data here: https://github.com/ben-aaron188/netanos/pull/7

After fixing these two items up, could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
