### 2017-06-25 2e02b2c
references to paper.bib + direct referencing of bibtex format in paper.md
