+++ paper/paper.bib
@@ -1,19 +1,20 @@
-@article{finkel,
+@article{finkel2005,
   author = {Finkel, J. R., Grenager, T. and Manning, C.},
   title = {Incorporating non-local information into information extraction systems by gibbs sampling},
   year = {2005},
   journal = {Proceedings of the 43rd annual meeting on association for computational linguistic},
-  pages = {363--370}
+  pages = {363--370},
+  doi = {10.3115/1219840.1219885}
 }
 
-@online{kelly,
+@online{kelly2016,
   author = {Kelly, S.},
   title = {NLP Compromise: Natural language processing in javascript},
   year = {2016},
   url = {https://github.com/nlpcompromise/compromise}
 }
 
-@online{srivastava,
+@online{srivastava2016,
   author = {Srivastava, N.},
   title = {ner: Client for Stanford Named Entity Reconginiton},
   year = {2016},
@@ -26,11 +27,12 @@
   url = {https://bitbucket.org/ukda/ukds.tools.textanonhelper/wiki/Home}
 }
 
-@article{warmelink,
+@article{warmelink2013,
   author = {Warmelink, L., Vrij, A., Mann, S., and Granhag, P. A},
   title = {Spatial and temporal details in intentions: A cue to detecting deception},
   journal = {Applied Cognitive Psychology},
   volume = {27},
   year = {2013},
-  pages = {101--106}
+  pages = {101--106},
+  doi = {10.1002/acp.2878}
 }
