### 2017-06-25 #293 @ben-aaron188
@arfon The references are updated and the pull request is merged. To avoid any versioning inaccuracies with npm, we’ve updated the npm release with these changes (which unfortunately implied that we needed to provide a new version in the `package.json`).
If possible, could you update the JOSS version to `v1.1.5`?

We’ve linked the `v1.1.5` tree in the master branch to Zenodo. Here’s the DOI 10.5281/zenodo.817892

The bagde is here (already included at the top of the readme):
[![DOI](https://zenodo.org/badge/86087510.svg)](https://zenodo.org/badge/latestdoi/86087510)
