### 2017-06-25 #293 @RichardLitt
Small note: thanks @ben-aaron188 for being so receptive to changes and fixes!
