### 2017-06-25 #293 @ben-aaron188
@RichardLitt Thanks for your helpful review and for taking the time with this tool! Much appreciated.

@arfon What a fantastic, smooth review process this is. I will recommend your journal to colleagues eager to put their open source research software out there! I think this kind of open reviewing will be the future of academic peer reviewing soon!
