### 2017-06-25 #293 @arfon
> @arfon What a fantastic, smooth review process this is. I will recommend your journal to colleagues eager to put their open source research software out there! I think this kind of open reviewing will be the future of academic peer reviewing soon!

💖 Thanks!
